Move extensions and utilities from to analyzer shared layer
