+++ src/Workspaces/SharedUtilitiesAndExtensions/Compiler/CSharp/Extensions/VariableDeclaratorExtensions.cs
