+++ src/Workspaces/SharedUtilitiesAndExtensions/Compiler/VisualBasic/Extensions/VariableDeclaratorSyntaxExtensions.vb
