+++ src/Workspaces/SharedUtilitiesAndExtensions/Workspace/VisualBasic/VisualBasicWorkspaceExtensions.projitems
@@ -11,7 +11,6 @@
   <ItemGroup>
     <Compile Include="$(MSBuildThisFileDirectory)Extensions\ArgumentListSyntaxExtensions.vb" />
     <Compile Include="$(MSBuildThisFileDirectory)Extensions\CallStatementSyntaxExtensions.vb" />
-    <Compile Include="$(MSBuildThisFileDirectory)Extensions\CastAnalyzer.vb" />
     <Compile Include="$(MSBuildThisFileDirectory)Extensions\CastExpressionSyntaxExtensions.vb" />
     <Compile Include="$(MSBuildThisFileDirectory)Extensions\CompilationUnitSyntaxExtensions.vb" />
     <Compile Include="$(MSBuildThisFileDirectory)Extensions\ContextQuery\IsStatementTerminatingTokenVisitor.vb" />
@@ -41,7 +40,6 @@
     <Compile Include="$(MSBuildThisFileDirectory)Extensions\SyntaxTriviaListExtensions.vb" />
     <Compile Include="$(MSBuildThisFileDirectory)Extensions\TypeBlockSyntaxExtensions.vb" />
     <Compile Include="$(MSBuildThisFileDirectory)Extensions\TypeSyntaxGeneratorVisitor.vb" />
-    <Compile Include="$(MSBuildThisFileDirectory)Extensions\VariableDeclaratorSyntaxExtensions.vb" />
     <Compile Include="$(MSBuildThisFileDirectory)LanguageServices\VisualBasicRemoveUnnecessaryImportsService.Rewriter.vb" />
     <Compile Include="$(MSBuildThisFileDirectory)LanguageServices\VisualBasicRemoveUnnecessaryImportsService.vb" />
     <Compile Include="$(MSBuildThisFileDirectory)LanguageServices\VisualBasicAddImportsService.vb" />
@@ -63,7 +61,6 @@
     <Compile Include="$(MSBuildThisFileDirectory)Utilities\NameSyntaxComparer.vb" />
     <Compile Include="$(MSBuildThisFileDirectory)Utilities\NameSyntaxIterator.vb" />
     <Compile Include="$(MSBuildThisFileDirectory)Utilities\PossibleDeclarationTypes.vb" />
-    <Compile Include="$(MSBuildThisFileDirectory)Utilities\SpeculationAnalyzer.vb" />
     <Compile Include="$(MSBuildThisFileDirectory)Utilities\SymbolExtensions.vb" />
     <Compile Include="$(MSBuildThisFileDirectory)Utilities\SyntaxKindSet.vb" />
     <Compile Include="$(MSBuildThisFileDirectory)Utilities\TokenComparer.vb" />
